feat: update system.rs

+++ src/modules/system.rs
@@ -20,10 +20,10 @@ impl SystemModule {
         let s = &mut app.system_state;
         s.cpu_usage = data.cpu_usage;
         s.cpu_cores = data.cpu_cores.to_vec();
-        s.mem_usage = data.mem_usage;
-        s.total_mem = data.total_mem;
-        s.swap_usage = data.swap_usage;
-        s.total_swap = data.total_swap;
+        s.mem_usage = data.mem_usage as f32;
+        s.total_mem = data.total_mem as f32;
+        s.swap_usage = data.swap_usage as f32;
+        s.total_swap = data.total_swap as f32;
         s.disks = data.disks;
         s.uptime = data.uptime;
         s.processes = data.processes;
